requirements.txt
# FastAPI Auth Service Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# gRPC (internal service-to-service auth)
grpcio>=1.60.0
grpcio-tools>=1.60.0

